requirements.txt
python-dotenv==1.0.1
scrapy==2.11.2
psycopg2==2.9.10
redis==5.1.1
pandas==2.2.3